move user logic to separate controller file
this completes the MVC architecture
